moved query builder to separate folders
