+++ mod.ts
@@ -1,4 +1,4 @@
-export * from "./src/Column.ts";
-export * from "./src/Schema.ts";
-export * from "./src/Table.ts";
-export * from "./src/TypeUtils.ts";
+export * from './clients/AbstractClient.ts'
+export * from './clients/ClientMySQL.ts'
+export * from './clients/ClientPostgreSQL.ts'
+export * from './clients/ClientSQLite.ts'
