+++ tests/query-builder/schema.pg.test.ts
@@ -1,23 +1,23 @@
-import { assertEquals } from "../deps.ts";
-import { Schema } from "../mod.ts";
+import { assertEquals } from "../../deps.ts";
+import { Schema } from "../../qb.ts";
 
 const strings = [
   {
     name: "Schema create",
     string: (() => {
-      const testSchema = new Schema("mysql");
+      const testSchema = new Schema();
       return testSchema.create("testTable", (table) => {
         table.id();
         table.timestamps();
       });
     })(),
     solution:
-      "CREATE TABLE testTable (id bigint AUTO_INCREMENT PRIMARY KEY, created_at timestamp (0) default current_timestamp, updated_at timestamp (0) default current_timestamp on update current_timestamp);",
+      "CREATE TABLE testTable (id bigserial PRIMARY KEY, created_at timestamp (0) default current_timestamp, updated_at timestamp (0) default current_timestamp); DROP TRIGGER IF EXISTS set_timestamp on public.testTable; CREATE TRIGGER set_timestamp BEFORE UPDATE ON public.testTable FOR EACH ROW EXECUTE PROCEDURE trigger_set_timestamp();",
   },
   {
     name: "Schema drop",
     string: (() => {
-      const testSchema = new Schema("mysql");
+      const testSchema = new Schema();
       return testSchema.drop("testTable");
     })(),
     solution: "DROP TABLE testTable;",
@@ -25,7 +25,7 @@ const strings = [
   {
     name: "Schema drop if exists",
     string: (() => {
-      const testSchema = new Schema("mysql");
+      const testSchema = new Schema();
       return testSchema.drop("testTable", true);
     })(),
     solution: "DROP TABLE IF EXISTS testTable;",
@@ -33,7 +33,7 @@ const strings = [
   {
     name: "Schema drop cascade",
     string: (() => {
-      const testSchema = new Schema("mysql");
+      const testSchema = new Schema();
       return testSchema.drop("testTable", false, true);
     })(),
     solution: "DROP TABLE testTable CASCADE;",
@@ -41,21 +41,21 @@ const strings = [
   {
     name: "Schema drop if exists cascade",
     string: (() => {
-      const testSchema = new Schema("mysql");
+      const testSchema = new Schema();
       return testSchema.drop("testTable", true, true);
     })(),
     solution: "DROP TABLE IF EXISTS testTable CASCADE;",
   },
   {
     name: "Schema hasTable",
-    string: new Schema("mysql").hasTable("testTable"),
-    solution: "show tables like 'testTable';",
+    string: new Schema().hasTable("testTable"),
+    solution: "SELECT to_regclass('testTable');",
   },
 ];
 
 strings.forEach(({ name, string, solution }) =>
   Deno.test({
-    name: "MySQL: " + (name || "Empty"),
+    name: "PG: " + (name || "Empty"),
     fn(): void {
       assertEquals(string, solution);
     },
